% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{antidiag}
\alias{antidiag}
\title{Antidiagonal matrix}
\usage{
antidiag(k)
}
\arguments{
\item{k}{The size of the matrix (kxk)}
}
\value{
A matrix of size kxk with 1 on the antidiagonal and 0 elsewhere
}
\description{
Antidiagonal matrix
}
\examples{
M<-antidiag(10)
print(M)

}
