% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_stage.R
\name{calibrate_finalDecision}
\alias{calibrate_finalDecision}
\title{Select the optimal critical value for the final analysis by targeting a specific FWER when no basket is truly active}
\usage{
calibrate_finalDecision(
  p_0 = 0.15,
  p_A = 0.45,
  K = 5,
  n = 12,
  lambda = 0.004,
  gamma = 1,
  FWER_target = 0.1,
  pi_critic_min = 0.1,
  pi_critic_max = 0.5,
  pi_critic_step = 0.05,
  nb_replicate = 10000,
  parallel.option = TRUE
)
}
\arguments{
\item{p_0}{the null probability of response, default is 0.15}

\item{p_A}{the active probability of response, default is 0.45}

\item{K}{the total number of baskets}

\item{n}{the sample size in each basket (give either a vector of length K or a single numeric value)}

\item{lambda}{the penalty parameter controlling strength of borrowing}

\item{gamma}{controls the weigths of L1 and L2 norms}

\item{FWER_target}{the Family Wise Error Rate targeted}

\item{pi_critic_min}{the minimum value considered in the grid of optimization for the critical value}

\item{pi_critic_max}{the maximum value considered in the grid of optimization for the critical value}

\item{pi_critic_step}{the steps length of the grid for the critical values}

\item{nb_replicate}{the number of simulated trials}

\item{parallel.option}{boolean, TRUE if simulations should be ran in parallel}
}
\value{
two numerical values; the first one pi_critic is the optimal critical values among the ones considered
the second value is the FWER obtained with this given critical value
}
\description{
Select the optimal critical value for the final analysis by targeting a specific FWER when no basket is truly active
}
