% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_stage.R
\name{computeMSE}
\alias{computeMSE}
\title{Compute average MSE under a given scenario of activity
for a grid of values for lambda and select the lambda that minimizes the average MSE}
\usage{
computeMSE(
  p_0 = 0.15,
  p_A = 0.45,
  A = 3,
  K = 5,
  n = 12,
  lambda_min = 0,
  lambda_max = 0.01,
  lambda_step = 0.001,
  gamma = 1,
  pi_critic = 0.35,
  nb_replicate = 10000
)
}
\arguments{
\item{p_0}{the null probability of response, default is 0.15}

\item{p_A}{the active probability of response, default is 0.45}

\item{A}{the number of active baskets}

\item{K}{the total number of baskets}

\item{n}{the sample size in each basket (give either a vector of length K or a single numeric value)}

\item{lambda_min}{minimum value considered for the grid of values of lambda}

\item{lambda_max}{maximum value considered for the grid of values of lambda}

\item{lambda_step}{the steps of the grid for of values for lambda}

\item{gamma}{controls the weigths of L1 and L2 norms}

\item{pi_critic}{the critical value of the final decision}

\item{nb_replicate}{the number of simulated trials}
}
\value{
the average MSE for each value of lambda and the lambda minimizing the average MSE
}
\description{
Compute average MSE under a given scenario of activity
for a grid of values for lambda and select the lambda that minimizes the average MSE
}
\examples{
computeMSE(nb_replicate = 100)

}
