% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logit_inv}
\alias{logit_inv}
\title{Inverse logit function}
\usage{
logit_inv(x)
}
\arguments{
\item{x}{a numeric value}
}
\value{
the inverse logit transformation of this number
}
\description{
Inverse logit function
}
\examples{
logit_inv(1)

}
