% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{matrixB}
\alias{matrixB}
\title{Creating the transformation matrix of the pairwise differences of the parameters}
\usage{
matrixB(K)
}
\arguments{
\item{K}{K is the number of baskets considered}
}
\value{
The transformation matrix of size (K*(K-1)/2+1) x K
}
\description{
Creating the transformation matrix of the pairwise differences of the parameters
}
\examples{
B<-matrixB(5)
print(B)

}
