% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_stage.R
\name{single_stage}
\alias{single_stage}
\title{Simulate, analyze and conclude on a simulated single-stage trial}
\usage{
single_stage(
  p_0 = 0.15,
  p_A = 0.45,
  A,
  K = 5,
  n = 12,
  lambda = 0.004,
  gamma = 1,
  pi_critic = 0.34
)
}
\arguments{
\item{p_0}{the null probability of response, default is 0.15}

\item{p_A}{the active probability of response, default is 0.45}

\item{A}{the number of active baskets}

\item{K}{the total number of baskets}

\item{n}{the sample size in each basket (give either a vector of length K or a single numeric value)}

\item{lambda}{the penalty parameter controlling strength of borrowing}

\item{gamma}{controls the weigths of L1 and L2 norms}

\item{pi_critic}{the critical value of the final decision}
}
\value{
outcomes is a vector of length K and each element is 1 if basket is declared active, 0 if else
}
\description{
Simulate, analyze and conclude on a simulated single-stage trial
}
\examples{
single_stage(A=3)

}
