% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stop_proba}
\alias{stop_proba}
\title{stop current execution if element is not between 0 and 1}
\usage{
stop_proba(x)
}
\arguments{
\item{x}{a numeric value}
}
\value{
an error message if x is not between 0 and 1
}
\description{
stop current execution if element is not between 0 and 1
}
\examples{
stop_proba(0.3)

}
