% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_stage.R
\name{summaryAllScenarios}
\alias{summaryAllScenarios}
\title{Give the probability of declaring drug activity in each basket under all scenarios of activity}
\usage{
summaryAllScenarios(
  p_0 = 0.15,
  p_A = 0.45,
  K = 5,
  n = 12,
  lambda = 0.004,
  gamma = 1,
  pi_critic = 0.34,
  nb_replicate = 10000
)
}
\arguments{
\item{p_0}{the null probability of response, default is 0.15}

\item{p_A}{the active probability of response, default is 0.45}

\item{K}{the total number of baskets}

\item{n}{the sample size in each basket (give either a vector of length K or a single numeric value)}

\item{lambda}{the penalty parameter}

\item{gamma}{controls the weigths of L1 and L2 norms}

\item{pi_critic}{the critical value of the final decision}

\item{nb_replicate}{the number of simulated trials}
}
\value{
a table of dimension (K+1)*K with the probability of rejecting the null
hypothesis in each basket under each scenario of activity A=0,...,K
and a vector of the Family Wise Error Rate for each scenario of activity
}
\description{
Give the probability of declaring drug activity in each basket under all scenarios of activity
}
\examples{
summaryAllScenarios(nb_replicate =1000, K=3, n=20)
}
