% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{emulate_trial}
\alias{emulate_trial}
\title{Emulate a single-stage trial}
\usage{
emulate_trial(p_0 = 0.15, p_A = 0.45, A, K = 5, n = 12)
}
\arguments{
\item{p_0}{the null probability of response, default is 0.15}

\item{p_A}{the active probability of response, default is 0.45}

\item{A}{the number of active baskets}

\item{K}{the total number of baskets}

\item{n}{the sample size in each basket (give either a vector of length K or a single numeric value)}
}
\value{
a vector of length sum(n) or n*K with all observations 0 or 1
}
\description{
Emulate a single-stage trial
}
\examples{
emulate_trial(A=3, K=6, n=c(12,12,10,8,5,11))

}
