% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{runFPmodel}
\alias{runFPmodel}
\title{Run the fusion-penalized logistic regression model}
\usage{
runFPmodel(x, y, lambda = 0.004, gamma = 1, pi_true = NULL)
}
\arguments{
\item{x}{is the covariates matrix of size sum(n)xK with element is 1 if observation belongs to the basket, 0 if else}

\item{y}{is the vector of observations, is 1 if individual responds to trt, 0 if not}

\item{lambda}{the penalty parameter controlling strength of borrowing}

\item{gamma}{controls the weigths of L1 and L2 norms}

\item{pi_true}{a vector of numeric values corresponding to the real values of the response probability in each basket,
each value should be between 0 and 1 and the length of the vector should be the same than the number of columns in x}
}
\value{
estimated values for the K response probabilities (natural scale) and
the squared difference between the estimates and the true value when a true value was specified
}
\description{
Run the fusion-penalized logistic regression model
}
\examples{
data_simu <- emulate_trial(A=3, K=6, n=c(12,12,10,8,5,11))
y <- data_simu$y
x <- data_simu$x
runFPmodel(x, y, lambda=0.008, gamma=1)

}
